refactor: admin config 와 client config 파일 분리
