+++ web/src/app/api/config/firebase.client.ts
@@ -1,10 +1,7 @@
-import { initializeApp, getApps } from 'firebase-admin/app';
-import { credential } from 'firebase-admin';
 import {
   getApps as clientGet,
   initializeApp as clientInit,
 } from 'firebase/app';
-import applicationDefault = credential.applicationDefault;
 
 const clientConfig = {
   apiKey: process.env.FIREBASE_API_KEY,
@@ -15,15 +12,8 @@ const clientConfig = {
   appId: process.env.FIREBASE_APP_ID,
   measurementId: process.env.FIREBASE_MEASUREMENT_ID,
 };
-const firebase = {
-  credential: applicationDefault(),
-};
 
-/**
- * 환경 설정 참고 : https://firebase.google.com/docs/admin/setup?hl=ko#windows
- */
-
-export const clientApp =
+const clientApp =
   clientGet().length === 0 ? clientInit(clientConfig) : clientGet()[0];
-export const firebaseApp =
-  getApps().length === 0 ? initializeApp(firebase) : getApps()[0];
+
+export default clientApp;
